# perf(rust, python): use binary abstraction in pow