--- a/crates/polars-arrow/src/kernels/mod.rs
+++ b/crates/polars-arrow/src/kernels/mod.rs
@@ -11,6 +11,7 @@ pub mod ewm;
 pub mod float;
 pub mod list;
 pub mod list_bytes_iter;
+pub mod pow;
 pub mod rolling;
 pub mod set;
 pub mod sort_partition;