--- a/crates/polars-plan/src/dsl/function_expr/pow.rs
+++ b/crates/polars-plan/src/dsl/function_expr/pow.rs
@@ -1,7 +1,8 @@
 use num::pow::Pow;
-use polars_arrow::utils::CustomIterTools;
+use polars_arrow::kernels::pow::pow as pow_kernel;
 use polars_core::export::num;
 use polars_core::export::num::{Float, ToPrimitive};
+use polars_core::utils::align_chunks_binary;
 
 use super::*;
 
@@ -63,15 +64,13 @@ where
             exponent.apply(|exp| Pow::pow(base, exp)).into_series(),
         ))
     } else {
+        let (ca_1, ca_2) = align_chunks_binary(base, exponent);
+        let chunks = ca_1
+            .downcast_iter()
+            .zip(ca_2.downcast_iter())
+            .map(|(arr_1, arr_2)| pow_kernel(arr_1, arr_2));
         Ok(Some(
-            base.into_iter()
-                .zip(exponent)
-                .map(|(opt_base, opt_exponent)| match (opt_base, opt_exponent) {
-                    (Some(base), Some(exponent)) => Some(num::pow::Pow::pow(base, exponent)),
-                    _ => None,
-                })
-                .collect_trusted::<ChunkedArray<T>>()
-                .into_series(),
+            ChunkedArray::from_chunk_iter(ca_1.name(), chunks).into_series(),
         ))
     }
 }