splitting this out from #10446 as it's gonna take me a bit longer to figure that one out

but this is simple and already seems to improve performance

timing
```python
size = 100_000_000
df = pl.DataFrame({'a': np.random.randint(0, 4, size=size), 'b': np.random.randn(size)}, schema_overrides={'a': pl.UInt8})
print(ipython.run_line_magic("timeit", "df.select(pl.col('a').pow(pl.col('a')))"))
print(ipython.run_line_magic("timeit", "df.select(pl.col('b').pow(pl.col('b')))"))
```
on kaggle I see:

- before: `8.35 s ± 2.41 s`, `4.42 s ± 21.2 ms`
- after: `7.29 s ± 2.3 s`, `2.92 s ± 18.6 ms`

source: https://www.kaggle.com/code/marcogorelli/polars-timing/notebook?scriptVersionId=140186516